+++ example/lib/main.dart
@@ -6,6 +6,7 @@ import 'routes/biometrics_page.dart';
 import 'routes/forgot_password_page.dart';
 import 'routes/home_page.dart';
 import 'routes/login_with_credentials_page.dart';
+import 'routes/login_with_custom_id_page.dart';
 import 'routes/manage_connections_page.dart';
 import 'routes/one_time_password_login_page.dart';
 import 'routes/register_with_email_page.dart';
@@ -59,6 +60,7 @@ class _MyAppState extends State<MyApp> {
         '/': (_) => HomePage(sdk: widget.sdk),
         '/send_request': (_) => SendRequestPage(sdk: widget.sdk),
         '/login_credentials': (_) => LoginWithCredentialsPage(sdk: widget.sdk),
+        '/login_custom_id': (_) => LoginWithCustomIdPage(sdk: widget.sdk),
         '/register_email': (_) => RegisterWithEmailPage(sdk: widget.sdk),
         '/account_information': (_) => AccountInformationPage(sdk: widget.sdk),
         '/manage_connections': (_) => ManageConnectionsPage(sdk: widget.sdk),